bakedbot_ai/bakedbot_ai/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic==1.8.2
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.5
psycopg2-binary==2.9.1
redis==3.5.3
celery==5.1.2
pinecone-client==2.0.11
boto3==1.18.44
scikit-learn==0.24.2
numpy==1.21.2
pytest==6.2.5
requests==2.26.0
